added code for chapter 07

+++ microservices/recommendation-service/src/main/java/com/pacifique/microservices/core/recommendation/services/RecommendationServiceImpl.java
@@ -1,6 +1,5 @@
 package com.pacifique.microservices.core.recommendation.services;
 
-import com.mongodb.DuplicateKeyException;
 import com.pacifique.microservices.api.core.recommendation.Recommendation;
 import com.pacifique.microservices.api.core.recommendation.RecommendationService;
 import com.pacifique.microservices.api.exceptions.InvalidInputException;
@@ -10,10 +9,12 @@ import com.pacifique.microservices.util.http.ServiceUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
 
-import java.util.ArrayList;
-import java.util.List;
+import static java.util.logging.Level.FINE;
 
 @RestController
 public class RecommendationServiceImpl implements RecommendationService {
@@ -31,38 +32,46 @@ public class RecommendationServiceImpl implements RecommendationService {
     }
 
     @Override
-    public Recommendation createRecommendation(Recommendation body) {
+    public Mono<Recommendation> createRecommendation(Recommendation body) {
+        if (body.getProductId() < 1) {
+            throw new InvalidInputException("Invalid productId: " + body.getProductId());
+        }
+
         RecommendationEntity entity = mapper.apiToEntity(body);
 
-        try {
-            entity = repository.save(entity);
-        }catch (DuplicateKeyException dke){
-            throw new InvalidInputException("Duplicate key, Product Id: " + body.getProductId() +", Recommendation Id: " + body.getRecommendationId());
-        }
-
-        LOG.debug("createRecommendation: created a recommendation entity: {}/{}", body.getProductId(), body.getRecommendationId());
-
-        return mapper.entityToApi(entity);
+        return repository.save(entity)
+                .log(LOG.getName(), FINE)
+                .onErrorMap(DuplicateKeyException.class,
+                        ex -> new InvalidInputException("Duplicate key, Product Id: " + body.getProductId() + ", Recommendation Id: " +
+                                body.getRecommendationId()))
+                .map(mapper::entityToApi);
     }
 
     @Override
-    public List<Recommendation> getRecommendations(int productId) {
+    public Flux<Recommendation> getRecommendations(int productId) {
         if (productId < 1) {
-            throw new  InvalidInputException("Invalid productId: "+productId);
+            throw new InvalidInputException("Invalid productId: " + productId);
         }
 
-        List<RecommendationEntity> entityList = repository.findByProductId(productId);
-        List<Recommendation> list = mapper.entityListToApiList(entityList);
-        list.forEach(e->e.setServiceAddress(serviceUtil.getServiceAddress()));
-
-        LOG.debug("getRecommendation: response size {}", list.size());
-
-        return list;
+        LOG.info("Will get recommendation for product with id: {}", productId);
+        return repository.findByProductId(productId)
+                .log(LOG.getName(), FINE)
+                .map(mapper::entityToApi)
+                .map(this::setServiceAddress);
     }
 
     @Override
-    public void deleteRecommendations(int productId) {
-    LOG.debug("deleteRecommendations: tries to delete recommendation for product with productId: {}", productId);
-    repository.deleteAll(repository.findByProductId(productId));
+    public Mono<Void> deleteRecommendations(int productId) {
+        if (productId < 1) {
+            throw new InvalidInputException("Invalid productId: " + productId);
+        }
+
+        LOG.debug("deleteRecommendations: tries to delete recommendation for product with productId: {}", productId);
+        return repository.deleteAll(repository.findByProductId(productId));
+    }
+
+    private Recommendation setServiceAddress(Recommendation e) {
+        e.setServiceAddress(serviceUtil.getServiceAddress());
+        return e;
     }
 }
